+++ lib/sbb_design_system_mobile.dart
@@ -18,6 +18,7 @@ export 'src/message/message.dart';
 export 'src/notification_box/notification_box.dart';
 export 'src/paginator/paginator.dart';
 export 'src/picker/sbb_picker.dart';
+export 'src/popup/popup.dart';
 export 'src/promotion_box/promotion_box.dart';
 export 'src/radio/radio.dart';
 export 'src/segmented_button/segmented_button.dart';