+++ lib/src/theme/sbb_theme.dart
@@ -12,10 +12,12 @@ import 'package:sbb_design_system_mobile/src/input/theme/default_sbb_text_input_
 import 'package:sbb_design_system_mobile/src/list_header/theme/default_sbb_list_header_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/message/theme/default_sbb_message_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/paginator/theme/default_sbb_paginator_theme_data.dart';
+import 'package:sbb_design_system_mobile/src/popup/theme/default_sbb_popup_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/radio/theme/default_sbb_radio_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/slider/theme/default_sbb_slider_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/stepper/theme/default_sbb_stepper_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/switch/theme/default_sbb_switch_theme_data.dart';
+import 'package:sbb_design_system_mobile/src/theme/styles/base_style_x.dart';
 import 'package:sbb_design_system_mobile/src/toast/theme/default_sbb_toast_theme_data.dart';
 
 import '../button/theme/default_button_themes.dart';
@@ -47,6 +49,7 @@ class SBBTheme {
     SBBListItemThemeData? listItemTheme,
     SBBMessageThemeData? messageTheme,
     SBBPaginatorThemeData? paginatorTheme,
+    SBBPopupThemeData? popupTheme,
     SBBRadioThemeData? radioTheme,
     SBBStatusThemeData? statusTheme,
     SBBSegmentedButtonThemeData? segmentedButtonTheme,
@@ -78,6 +81,7 @@ class SBBTheme {
     messageTheme: messageTheme,
     paginatorTheme: paginatorTheme,
     radioTheme: radioTheme,
+    popupTheme: popupTheme,
     statusTheme: statusTheme,
     segmentedButtonTheme: segmentedButtonTheme,
     sliderTheme: sliderTheme,
@@ -107,6 +111,7 @@ class SBBTheme {
     SBBListItemThemeData? listItemTheme,
     SBBMessageThemeData? messageTheme,
     SBBPaginatorThemeData? paginatorTheme,
+    SBBPopupThemeData? popupTheme,
     SBBRadioThemeData? radioTheme,
     SBBStatusThemeData? statusTheme,
     SBBSegmentedButtonThemeData? segmentedButtonTheme,
@@ -136,6 +141,7 @@ class SBBTheme {
     listItemTheme: listItemTheme,
     messageTheme: messageTheme,
     paginatorTheme: paginatorTheme,
+    popupTheme: popupTheme,
     radioTheme: radioTheme,
     statusTheme: statusTheme,
     stepperTheme: stepperTheme,
@@ -169,6 +175,7 @@ class SBBTheme {
     SBBListItemThemeData? listItemTheme,
     SBBMessageThemeData? messageTheme,
     SBBPaginatorThemeData? paginatorTheme,
+    SBBPopupThemeData? popupTheme,
     SBBRadioThemeData? radioTheme,
     SBBSegmentedButtonThemeData? segmentedButtonTheme,
     SBBSliderThemeData? sliderTheme,
@@ -230,6 +237,9 @@ class SBBTheme {
     final defaultPaginatorTheme = DefaultSBBPaginatorThemeData(mergedBaseStyle);
     final mergedPaginatorTheme = defaultPaginatorTheme.merge(paginatorTheme);
 
+    final defaultPopupTheme = DefaultSBBPopupThemeData(mergedBaseStyle);
+    final mergedPopupTheme = defaultPopupTheme.merge(popupTheme);
+
     final defaultRadioTheme = DefaultSBBRadioThemeData(mergedBaseStyle);
     final mergedRadioTheme = defaultRadioTheme.merge(radioTheme);
 
@@ -280,6 +290,7 @@ class SBBTheme {
       listItemTheme: mergedListItemTheme,
       messageTheme: mergedMessageTheme,
       paginatorTheme: mergedPaginatorTheme,
+      popupTheme: mergedPopupTheme,
       radioTheme: mergedRadioTheme,
       statusTheme: mergedStatusTheme,
       segmentedButtonTheme: mergedSegmentedButtonTheme,
@@ -312,6 +323,7 @@ class SBBTheme {
     required SBBListItemThemeData listItemTheme,
     required SBBMessageThemeData messageTheme,
     required SBBPaginatorThemeData paginatorTheme,
+    required SBBPopupThemeData popupTheme,
     required SBBRadioThemeData radioTheme,
     required SBBStatusThemeData statusTheme,
     required SBBSegmentedButtonThemeData segmentedButtonTheme,
@@ -361,6 +373,7 @@ class SBBTheme {
         listItemTheme,
         messageTheme,
         paginatorTheme,
+        popupTheme,
         radioTheme,
         statusTheme,
         segmentedButtonTheme,