+++ CHANGELOG.md
@@ -30,6 +30,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - BREAKING: overhauled `SBBMessage` component - see [v5 migration guide]
 - BREAKING: overhauled `SBBModal` component and renamed to `SBBBottomSheet` - see [v5 migration guide]
 - BREAKING: overhauled `SBBPagination` and renamed to `SBBPaginator` - see [v5 migration guide]
+- BREAKING: overhauled `SBBModalPopup` and renamed to `SBBPopup` - see [v5 migration guide]
 - BREAKING: overhauled `SBBRadio` with `SBBRadioGroup` ancestor - see [v5 migration guide]
 - BREAKING: overhauled `SBBSegmentedButton` component - see [v5 migration guide]
 - BREAKING: overhauled `SBBSlider` component - see [v5 migration guide]
@@ -39,6 +40,9 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - BREAKING: different theming / styling for `SBBSwitch` - see [v5 migration guide]
 - BREAKING: overhauled `SBBTextField` and renamed to `SBBTextInput` - see [v5 migration guide]
 - BREAKING: removed default left margin on component `SBBTextInput` (#402)
+- renamed TextStylesExtensions to SBBTextStyleExtensions
+- BREAKING: renamed `ContractionState` to `SBBContractionState`
+- BREAKING: renamed `ContractibleState` to `SBBContractibleState`
 
 ### Removed
 
@@ -52,8 +56,9 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - Removed deprecated `SBBGroup`
 - Removed deprecated `SBBOnboarding`
 - Removed deprecated `sbbDefaultSpacing`
+- Removed `CascadeColumnParentData` as it is not directly used by clients
 - Removed `SBBTextField` - replaced by `SBBTextInput`
-
+- Removed `merge` extension on `SBBBaseStyle`
 
 ## [4.8.0] - 2026-03-05